backend/requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-groq
langchain-chroma
sentence-transformers
langgraph
pydantic
chromadb
tqdm